requirements.txt
discord.py==2.5.2
apscheduler==3.10.4